spawning model refinement and spawning training data

+++ data-raw/scripts/hydraulic-training-data-vector.R
@@ -57,7 +57,7 @@ stan_hsi_by_flow <-
   vector_import_srh2d(units = "m") |>
   vector_calculate_hsi(hsi_func = vector_dvhsi_hqt)
 
-###
+### ORIGINAL VERSION
 
 stan_mesh_prepped <-
   stan_thiessen |>
@@ -73,7 +73,7 @@ outpath <- here::here("data-raw", "results", "fsa_stan.Rds")
 
 stan_result |> saveRDS(outpath)
 
-###
+### VARIATION FOR REARING - BASEFLOW REMOVAL
 
 # define baseflow mask as 500 cfs (COMID 2819818 observed dry season baseflow is 306 cfs)
 stan_baseflow_poly <- stan_thiessen |>
@@ -95,6 +95,21 @@ stan_result_nbfc <-
 outpath <- here::here("data-raw", "results", "fsa_stan_nbfc.Rds")
 
 stan_result_nbfc |> saveRDS(outpath)
+
+### VARIATION FOR SPAWNING - ALTERNATIVE HSI
+
+stan_hsi_by_flow_spawning <-
+  stan_csvs |>
+  vector_import_srh2d(units = "m") |>
+  vector_calculate_hsi(hsi_func = vector_dvhsi_spawning)
+
+stan_result_spawning <-
+  vector_summarize_hsi(stan_mesh_prepped, stan_hsi_by_flow_spawning) |>
+  suitability_postprocess(stan_comid, .group_var = comid)
+
+outpath <- here::here("data-raw", "results", "fsa_stan_spawning.Rds")
+
+stan_result_spawning |> saveRDS(outpath)
 
 
 # YUBA RIVER -------------------------------------------------------------
@@ -155,24 +170,32 @@ yuba_mesh_data_by_reach <-
          geom_thiessen = future_map(filename_thiessen,
                              function(x) st_read(x, as_tibble=T)))
 
-###
+### ORIGINAL VERSION
 
-yuba_mesh_prepped_by_reach <-
-  yuba_mesh_data_by_reach |>
-  mutate(mesh_prepped = future_pmap(list(geom_vertices, geom_thiessen, reach),
-                             function(vtx, thp, rch) {
-                               thp |>
-                                 st_join(vtx |>
-                                           transmute(vid=row_number()),
-                                         join=st_contains_properly, left = FALSE) |> #st_nearest_feature) |>
-                                 arrange(vid) |>
-                                 vector_prep_mesh(group_polygons = yuba_comid |> filter(reach==rch),
-                                                  .group_var = comid,
-                                                  .id_var = vid)
-                             })) |>
-  select(reach, mesh_prepped)
+if(!file.exists(file.path(yuba_dir, "yuba_mesh_prepped_by_reach.Rds"))) {
 
-yuba_mesh_prepped_by_reach |> saveRDS(file.path(yuba_dir, "yuba_mesh_prepped_by_reach.Rds"))
+  yuba_mesh_prepped_by_reach <-
+    yuba_mesh_data_by_reach |>
+    mutate(mesh_prepped = future_pmap(list(geom_vertices, geom_thiessen, reach),
+                               function(vtx, thp, rch) {
+                                 thp |>
+                                   st_join(vtx |>
+                                             transmute(vid=row_number()),
+                                           join=st_contains_properly, left = FALSE) |> #st_nearest_feature) |>
+                                   arrange(vid) |>
+                                   vector_prep_mesh(group_polygons = yuba_comid |> filter(reach==rch),
+                                                    .group_var = comid,
+                                                    .id_var = vid)
+                               })) |>
+    select(reach, mesh_prepped)
+
+  yuba_mesh_prepped_by_reach |> saveRDS(file.path(yuba_dir, "yuba_mesh_prepped_by_reach.Rds"))
+
+} else {
+
+  yuba_mesh_prepped_by_reach <- readRDS(file.path(yuba_dir, "yuba_mesh_prepped_by_reach.Rds"))
+
+}
 
 yuba_output_by_reach <-
   inner_join(yuba_hsi_by_flow_by_reach, yuba_mesh_prepped_by_reach,
@@ -194,7 +217,7 @@ outpath <- here::here("data-raw", "results", "fsa_yuba.Rds")
 
 yuba_result |> saveRDS(outpath)
 
-###
+### VARIATION FOR REARING - BASEFLOW REMOVAL
 
 yuba_baseflow_poly <-
   yuba_mesh_data_by_reach |>
@@ -250,3 +273,32 @@ outpath <- here::here("data-raw", "results", "fsa_yuba_nbfc.Rds")
 
 yuba_result_nbfc |> saveRDS(outpath)
 
+### VARIATION FOR SPAWNING - ALTERNATIVE HSI
+
+yuba_hsi_by_flow_by_reach_spawning <-
+  yuba_srh2d_result_by_reach |>
+  mutate(tbl_hsi = future_map(tbl_results,
+                              function(x) vector_calculate_hsi(x, hsi_func = vector_dvhsi_spawning))) |>
+  select(reach, tbl_hsi)
+
+yuba_hsi_by_flow_by_reach_spawning |> saveRDS(file.path(yuba_dir, "yuba_hsi_by_flow_by_reach_spawning.Rds"))
+
+yuba_output_by_reach_spawning <-
+  inner_join(yuba_hsi_by_flow_by_reach_spawning, yuba_mesh_prepped_by_reach,
+             by=join_by(reach), relationship="one-to-one") |>
+  mutate(hsi_output = pmap(list(mesh_prepped, tbl_hsi),
+                           function(x, y) vector_summarize_hsi(x, y, .group_var=comid, .id_var=vid))) |>
+  mutate(hsi_final = pmap(list(hsi_output, reach),
+                          function(x, y) suitability_postprocess(x, yuba_comid |> filter(reach==y), .group_var=comid))) |>
+  select(reach, hsi_final)
+
+yuba_output_by_reach_spawning |> saveRDS(file.path(yuba_dir, "yuba_output_by_reach_spawning.Rds"))
+
+yuba_result_spawning <-
+  yuba_output_by_reach_spawning |>
+  select(reach, hsi_final) |>
+  unnest(hsi_final)
+
+outpath <- here::here("data-raw", "results", "fsa_yuba_spawning.Rds")
+
+yuba_result_spawning |> saveRDS(outpath)
